+++ dbt_platform_helper/utils/validation.py
@@ -6,13 +6,12 @@ import click
 import yaml
 from schema import SchemaError
 from yaml.parser import ParserError
-from yamllint import config
-from yamllint import linter
 
 from dbt_platform_helper.constants import CODEBASE_PIPELINES_KEY
 from dbt_platform_helper.constants import ENVIRONMENTS_KEY
 from dbt_platform_helper.constants import PLATFORM_CONFIG_FILE
 from dbt_platform_helper.constants import PLATFORM_HELPER_VERSION_FILE
+from dbt_platform_helper.providers.config import ConfigProvider
 from dbt_platform_helper.providers.platform_config_schema import PlatformConfigSchema
 from dbt_platform_helper.utils.aws import get_supported_opensearch_versions
 from dbt_platform_helper.utils.aws import get_supported_redis_versions
@@ -275,31 +274,12 @@ def _validate_environment_pipelines_triggers(config):
         abort_with_error(error_message + "\n  ".join(errors))
 
 
-def lint_yaml_for_duplicate_keys(file_path):
-    lint_yaml_config = """
-rules:
-  key-duplicates: enable
-"""
-    yaml_config = config.YamlLintConfig(lint_yaml_config)
-
-    with open(file_path, "r") as yaml_file:
-        file_contents = yaml_file.read()
-        results = linter.run(file_contents, yaml_config)
-
-    parsed_results = [
-        "\t" + f"Line {result.line}: {result.message}".replace(" in mapping (key-duplicates)", "")
-        for result in results
-    ]
-
-    return parsed_results
-
-
 def load_and_validate_platform_config(path=PLATFORM_CONFIG_FILE, disable_file_check=False):
     if not disable_file_check:
         config_file_check(path)
     try:
         conf = yaml.safe_load(Path(path).read_text())
-        duplicate_keys = lint_yaml_for_duplicate_keys(path)
+        duplicate_keys = ConfigProvider.lint_yaml_for_duplicate_keys(path)
         if duplicate_keys:
             abort_with_error(
                 "Duplicate keys found in platform-config:"
